Change type for time

+++ lib/logging/log.dart
@@ -1,6 +1,6 @@
 abstract class Log {
   void startTrace(String trace);
   void endTrace(String trace);
-  void addSample(String key, double milliseconds);
+  void addSample(String key, int milliseconds);
   void addEvent(String event, {String? payload});
 }
